requirements.txt
# Snaphire Analytics Platform - Final Requirements

# Core Framework
streamlit>=1.30.0
flask>=3.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
xlrd>=2.0.0
xlsxwriter>=3.1.0

# Database & Config
pymongo>=4.6.0
python-dotenv>=1.0.0

# Authentication & Security
bcrypt>=4.1.0

# PDF & Reconciliation Utilities
pdfplumber>=0.10.0
PyPDF2>=3.0.0
azure-ai-documentintelligence>=1.0.0b1
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.23.0

# Networking & Production
requests>=2.31.0
gunicorn>=21.2.0
